Added support for custom Vega charts.
Function Name Changes:
renderVegaChart() -> renderVegaLiteChart()

Function Additions:
renderVegaChart()

Implementation Notes:
Autosizing a rendered Vega chart based on the css container dimensions is achieved by calling the Vega function 'containerSize()' which will return an array [width, height]. The chart is initialized to the dimensions and then update on any window resize events. If this part of the vega spec is missing when rendered, it will mess up how the dashboard grid is displayed. I have not checked to see if a user can accidentally overwrite this if they submit their own custom signal spec.